refactor: split migration maintenance helpers
